+++ src/api/requests/global-banners/getGlobalBanners.ts
