+++ src/mock-server/handlers/global-banners/fixtures/global-banner-information.ts
@@ -1,4 +1,4 @@
-import { ResponseSchema } from '@/api/requests/global-banners/handlers'
+import { ResponseSchema } from '@/api/requests/global-banners/getGlobalBanners'
 
 export const globalBannerInformation: ResponseSchema = {
   'active-global-banner': {
